dvf-producer/requirements.txt
kafka-python-ng==2.2.0
requests==2.31.0
six==1.16.0
SQLAlchemy>=1.4
psycopg2-binary>=2.9